+++ iop/hdd/pfs/src/pfs_fioctl.h
@@ -15,8 +15,8 @@
 //	Function declarations
 
 int pfsFioIoctl(iop_file_t *f, int cmd, void *param);
-int pfsFioIoctl2(iop_file_t *f, int cmd, void *arg, size_t arglen, void *buf, size_t buflen);
-int pfsFioDevctl(iop_file_t *f, const char *name, int cmd, void *arg, size_t arglen, void *buf, size_t buflen);
+int pfsFioIoctl2(iop_file_t *f, int cmd, void *arg, unsigned int arglen, void *buf, unsigned int buflen);
+int pfsFioDevctl(iop_file_t *f, const char *name, int cmd, void *arg, unsigned int arglen, void *buf, unsigned int buflen);
 
 void pfsFioDevctlCloseAll(void);
